+++ src/RcppExports.cpp
@@ -51,15 +51,15 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// ssm_parameters
-arma::vec ssm_parameters(arma::vec scores, arma::vec angles);
-RcppExport SEXP _circumplex_ssm_parameters(SEXP scoresSEXP, SEXP anglesSEXP) {
+// ssm_parameters_cpp
+arma::vec ssm_parameters_cpp(arma::vec scores, arma::vec angles);
+RcppExport SEXP _circumplex_ssm_parameters_cpp(SEXP scoresSEXP, SEXP anglesSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< arma::vec >::type scores(scoresSEXP);
     Rcpp::traits::input_parameter< arma::vec >::type angles(anglesSEXP);
-    rcpp_result_gen = Rcpp::wrap(ssm_parameters(scores, angles));
+    rcpp_result_gen = Rcpp::wrap(ssm_parameters_cpp(scores, angles));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -131,7 +131,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_circumplex_angle_dev", (DL_FUNC) &_circumplex_angle_dev, 2},
     {"_circumplex_angle_median", (DL_FUNC) &_circumplex_angle_median, 1},
     {"_circumplex_compare_pi", (DL_FUNC) &_circumplex_compare_pi, 1},
-    {"_circumplex_ssm_parameters", (DL_FUNC) &_circumplex_ssm_parameters, 2},
+    {"_circumplex_ssm_parameters_cpp", (DL_FUNC) &_circumplex_ssm_parameters_cpp, 2},
     {"_circumplex_group_parameters", (DL_FUNC) &_circumplex_group_parameters, 2},
     {"_circumplex_col_means", (DL_FUNC) &_circumplex_col_means, 1},
     {"_circumplex_mean_scores", (DL_FUNC) &_circumplex_mean_scores, 3},